requirement.txt
appdirs==1.4.3
APScheduler==3.6.3
asyncio==3.4.3
attrs==19.3.0
Automat==0.8.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
DBUtils==1.3
fake-useragent==0.1.11
fonttools==4.2.2
gne==0.1.7
HTMLParser==0.0.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jieba==0.40
jsonpath==0.82
lxml==4.4.1
numpy==1.18.0
opencv-python==4.1.2.30
pandas==0.25.3
parsel==1.5.2
Pillow==6.2.1
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybind11==2.4.3
pycparser==2.19
PyDispatcher==2.0.5
pyee==7.0.1
PyExecJS==1.5.1
PyHamcrest==1.9.0
pymssql==2.1.4
pynput==1.6.3
pyOpenSSL==19.1.0
pyppeteer==0.0.25
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
queuelib==1.5.0
requests==2.22.0
retrying==1.3.3
Scrapy==1.8.0
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
SQLAlchemy==1.3.12
text-unidecode==1.3
tqdm==4.42.1
Twisted==19.10.0
tzlocal==2.0.0
urllib3==1.25.6
w3lib==1.21.0
websockets==8.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.7.1
